chore: export basic types
